+++ icu4j/main/classes/core/src/com/ibm/icu/impl/breakiter/BytesDictionaryMatcher.java
@@ -6,11 +6,12 @@
  * others. All Rights Reserved.                                                *
  *******************************************************************************
  */
-package com.ibm.icu.text;
+package com.ibm.icu.impl.breakiter;
 
 import java.text.CharacterIterator;
 
 import com.ibm.icu.impl.Assert;
+import com.ibm.icu.text.UCharacterIterator;
 import com.ibm.icu.util.BytesTrie;
 import com.ibm.icu.util.BytesTrie.Result;
