+++ icu4j/main/classes/core/src/com/ibm/icu/impl/breakiter/DictionaryBreakEngine.java
@@ -6,13 +6,14 @@
  * others. All Rights Reserved.                                                *
  *******************************************************************************
  */
-package com.ibm.icu.text;
+package com.ibm.icu.impl.breakiter;
 
 import java.text.CharacterIterator;
 
 import com.ibm.icu.impl.CharacterIteration;
+import com.ibm.icu.text.UnicodeSet;
 
-abstract class DictionaryBreakEngine implements LanguageBreakEngine {
+public abstract class DictionaryBreakEngine implements LanguageBreakEngine {
 
     /* Helper class for improving readability of the Thai/Lao/Khmer word break
      * algorithm.
@@ -88,7 +89,7 @@ abstract class DictionaryBreakEngine implements LanguageBreakEngine {
      *  For internal use only.
      * @internal
      */
-    static class DequeI implements Cloneable {
+    public static class DequeI implements Cloneable {
         private int[] data = new int[50];
         private int lastIdx = 4;   // or base of stack. Index of element.
         private int firstIdx = 4;  // or Top of Stack. Index of element + 1.
@@ -100,11 +101,11 @@ abstract class DictionaryBreakEngine implements LanguageBreakEngine {
             return result;
         }
 
-        int size() {
+        public int size() {
             return firstIdx - lastIdx;
         }
 
-        boolean isEmpty() {
+        public boolean isEmpty() {
             return size() == 0;
         }
 
@@ -114,26 +115,26 @@ abstract class DictionaryBreakEngine implements LanguageBreakEngine {
             data = newData;
         }
 
-        void offer(int v) {
+        public void offer(int v) {
             // Note that the actual use cases of offer() add at most one element.
             //   We make no attempt to handle more than a few.
             assert lastIdx > 0;
             data[--lastIdx] = v;
         }
 
-        void push(int v) {
+        public void push(int v) {
             if (firstIdx >= data.length) {
                 grow();
             }
             data[firstIdx++] = v;
         }
 
-        int pop() {
+        public int pop() {
             assert size() > 0;
             return data[--firstIdx];
         }
 
-        int peek() {
+        public int peek() {
             assert size() > 0;
             return data[firstIdx - 1];
         }
@@ -157,12 +158,12 @@ abstract class DictionaryBreakEngine implements LanguageBreakEngine {
             return false;
         }
 
-        int elementAt(int i) {
+        public int elementAt(int i) {
             assert i < size();
             return data[lastIdx + i];
         }
 
-        void removeAllElements() {
+        public void removeAllElements() {
             lastIdx = firstIdx = 4;
         }
     }
