+++ icu4j/main/classes/core/src/com/ibm/icu/impl/breakiter/DictionaryMatcher.java
@@ -6,7 +6,7 @@
  * others. All Rights Reserved.                                                *
  *******************************************************************************
  */
-package com.ibm.icu.text;
+package com.ibm.icu.impl.breakiter;
 
 import java.text.CharacterIterator;
