+++ icu4j/main/classes/core/src/com/ibm/icu/impl/breakiter/KhmerBreakEngine.java
@@ -6,7 +6,7 @@
  * others. All Rights Reserved.                                                *
  *******************************************************************************
  */
-package com.ibm.icu.text;
+package com.ibm.icu.impl.breakiter;
 
 import java.io.IOException;
 import java.text.CharacterIterator;
@@ -14,42 +14,44 @@ import java.text.CharacterIterator;
 import com.ibm.icu.lang.UCharacter;
 import com.ibm.icu.lang.UProperty;
 import com.ibm.icu.lang.UScript;
+import com.ibm.icu.text.UnicodeSet;
 
-class LaoBreakEngine extends DictionaryBreakEngine {
+public class KhmerBreakEngine extends DictionaryBreakEngine {
 
-    // Constants for LaoBreakIterator
+    // Constants for KhmerBreakIterator
     // How many words in a row are "good enough"?
-    private static final byte LAO_LOOKAHEAD = 3;
+    private static final byte KHMER_LOOKAHEAD = 3;
     // Will not combine a non-word with a preceding dictionary word longer than this
-    private static final byte LAO_ROOT_COMBINE_THRESHOLD = 3;
+    private static final byte KHMER_ROOT_COMBINE_THRESHOLD = 3;
     // Will not combine a non-word that shares at least this much prefix with a
     // dictionary word with a preceding word
-    private static final byte LAO_PREFIX_COMBINE_THRESHOLD = 3;
+    private static final byte KHMER_PREFIX_COMBINE_THRESHOLD = 3;
     // Minimum word size
-    private static final byte LAO_MIN_WORD = 2;
+    private static final byte KHMER_MIN_WORD = 2;
+    // Minimum number of characters for two words
+    private static final byte KHMER_MIN_WORD_SPAN = KHMER_MIN_WORD * 2;
+
 
     private DictionaryMatcher fDictionary;
-    private static UnicodeSet fLaoWordSet;
+    private static UnicodeSet fKhmerWordSet;
     private static UnicodeSet fEndWordSet;
     private static UnicodeSet fBeginWordSet;
     private static UnicodeSet fMarkSet;
 
     static {
         // Initialize UnicodeSets
-        fLaoWordSet = new UnicodeSet();
+        fKhmerWordSet = new UnicodeSet();
         fMarkSet = new UnicodeSet();
         fBeginWordSet = new UnicodeSet();
 
-        fLaoWordSet.applyPattern("[[:Laoo:]&[:LineBreak=SA:]]");
-        fLaoWordSet.compact();
+        fKhmerWordSet.applyPattern("[[:Khmer:]&[:LineBreak=SA:]]");
+        fKhmerWordSet.compact();
 
-        fMarkSet.applyPattern("[[:Laoo:]&[:LineBreak=SA:]&[:M:]]");
+        fMarkSet.applyPattern("[[:Khmer:]&[:LineBreak=SA:]&[:M:]]");
         fMarkSet.add(0x0020);
-        fEndWordSet = new UnicodeSet(fLaoWordSet);
-        fEndWordSet.remove(0x0EC0, 0x0EC4); // prefix vowels
-        fBeginWordSet.add(0x0E81, 0x0EAE); // basic consonants (including holes for corresponding Thai characters)
-        fBeginWordSet.add(0x0EDC, 0x0EDD); // digraph consonants (no Thai equivalent)
-        fBeginWordSet.add(0x0EC0, 0x0EC4); // prefix vowels
+        fEndWordSet = new UnicodeSet(fKhmerWordSet);
+        fBeginWordSet.add(0x1780, 0x17B3);
+        fEndWordSet.remove(0x17D2); // KHMER SIGN COENG that combines some following characters
 
         // Compact for caching
         fMarkSet.compact();
@@ -57,23 +59,23 @@ class LaoBreakEngine extends DictionaryBreakEngine {
         fBeginWordSet.compact();
 
         // Freeze the static UnicodeSet
-        fLaoWordSet.freeze();
+        fKhmerWordSet.freeze();
         fMarkSet.freeze();
         fEndWordSet.freeze();
         fBeginWordSet.freeze();
     }
 
-    public LaoBreakEngine() throws IOException {
-        setCharacters(fLaoWordSet);
+    public KhmerBreakEngine() throws IOException {
+        setCharacters(fKhmerWordSet);
         // Initialize dictionary
-        fDictionary = DictionaryData.loadDictionaryFor("Laoo");
+        fDictionary = DictionaryData.loadDictionaryFor("Khmr");
     }
 
     @Override
     public boolean equals(Object obj) {
         // Normally is a singleton, but it's possible to have duplicates
         //   during initialization. All are equivalent.
-        return obj instanceof LaoBreakEngine;
+        return obj instanceof KhmerBreakEngine;
     }
 
     @Override
@@ -84,36 +86,36 @@ class LaoBreakEngine extends DictionaryBreakEngine {
     @Override
     public boolean handles(int c) {
         int script = UCharacter.getIntPropertyValue(c, UProperty.SCRIPT);
-        return (script == UScript.LAO);
+        return (script == UScript.KHMER);
     }
 
     @Override
     public int divideUpDictionaryRange(CharacterIterator fIter, int rangeStart, int rangeEnd,
             DequeI foundBreaks) {
 
-
-        if ((rangeEnd - rangeStart) < LAO_MIN_WORD) {
+        if ((rangeEnd - rangeStart) < KHMER_MIN_WORD_SPAN) {
             return 0;  // Not enough characters for word
         }
         int wordsFound = 0;
         int wordLength;
         int current;
-        PossibleWord words[] = new PossibleWord[LAO_LOOKAHEAD];
-        for (int i = 0; i < LAO_LOOKAHEAD; i++) {
+        PossibleWord words[] = new PossibleWord[KHMER_LOOKAHEAD];
+        for (int i = 0; i < KHMER_LOOKAHEAD; i++) {
             words[i] = new PossibleWord();
         }
         int uc;
 
         fIter.setIndex(rangeStart);
+
         while ((current = fIter.getIndex()) < rangeEnd) {
             wordLength = 0;
 
             //Look for candidate words at the current position
-            int candidates = words[wordsFound%LAO_LOOKAHEAD].candidates(fIter, fDictionary, rangeEnd);
+            int candidates = words[wordsFound % KHMER_LOOKAHEAD].candidates(fIter, fDictionary, rangeEnd);
 
             // If we found exactly one, use that
             if (candidates == 1) {
-                wordLength = words[wordsFound%LAO_LOOKAHEAD].acceptMarked(fIter);
+                wordLength = words[wordsFound % KHMER_LOOKAHEAD].acceptMarked(fIter);
                 wordsFound += 1;
             }
 
@@ -123,9 +125,9 @@ class LaoBreakEngine extends DictionaryBreakEngine {
                 // If we're already at the end of the range, we're done
                 if (fIter.getIndex() < rangeEnd) {
                     do {
-                        if (words[(wordsFound+1)%LAO_LOOKAHEAD].candidates(fIter, fDictionary, rangeEnd) > 0) {
+                        if (words[(wordsFound+1)%KHMER_LOOKAHEAD].candidates(fIter, fDictionary, rangeEnd) > 0) {
                             // Followed by another dictionary word; mark first word as a good candidate
-                            words[wordsFound%LAO_LOOKAHEAD].markCurrent();
+                            words[wordsFound%KHMER_LOOKAHEAD].markCurrent();
 
                             // If we're already at the end of the range, we're done
                             if (fIter.getIndex() >= rangeEnd) {
@@ -135,16 +137,16 @@ class LaoBreakEngine extends DictionaryBreakEngine {
                             // See if any of the possible second words is followed by a third word
                             do {
                                 // If we find a third word, stop right away
-                                if (words[(wordsFound+2)%LAO_LOOKAHEAD].candidates(fIter, fDictionary, rangeEnd) > 0) {
-                                    words[wordsFound%LAO_LOOKAHEAD].markCurrent();
+                                if (words[(wordsFound+2)%KHMER_LOOKAHEAD].candidates(fIter, fDictionary, rangeEnd) > 0) {
+                                    words[wordsFound%KHMER_LOOKAHEAD].markCurrent();
                                     foundBest = true;
                                     break;
                                 }
-                            } while (words[(wordsFound+1)%LAO_LOOKAHEAD].backUp(fIter));
+                            } while (words[(wordsFound+1)%KHMER_LOOKAHEAD].backUp(fIter));
                         }
-                    } while (words[wordsFound%LAO_LOOKAHEAD].backUp(fIter) && !foundBest);
+                    } while (words[wordsFound%KHMER_LOOKAHEAD].backUp(fIter) && !foundBest);
                 }
-                wordLength = words[wordsFound%LAO_LOOKAHEAD].acceptMarked(fIter);
+                wordLength = words[wordsFound%KHMER_LOOKAHEAD].acceptMarked(fIter);
                 wordsFound += 1;
             }
 
@@ -153,13 +155,13 @@ class LaoBreakEngine extends DictionaryBreakEngine {
             // just found (if there is one), but only if the preceding word does not exceed
             // the threshold.
             // The text iterator should now be positioned at the end of the word we found.
-            if (fIter.getIndex() < rangeEnd && wordLength < LAO_ROOT_COMBINE_THRESHOLD) {
+            if (fIter.getIndex() < rangeEnd && wordLength < KHMER_ROOT_COMBINE_THRESHOLD) {
                 // If it is a dictionary word, do nothing. If it isn't, then if there is
                 // no preceding word, or the non-word shares less than the minimum threshold
                 // of characters with a dictionary word, then scan to resynchronize
-                if (words[wordsFound%LAO_LOOKAHEAD].candidates(fIter, fDictionary, rangeEnd) <= 0 &&
+                if (words[wordsFound%KHMER_LOOKAHEAD].candidates(fIter, fDictionary, rangeEnd) <= 0 &&
                         (wordLength == 0 ||
-                                words[wordsFound%LAO_LOOKAHEAD].longestPrefix() < LAO_PREFIX_COMBINE_THRESHOLD)) {
+                                words[wordsFound%KHMER_LOOKAHEAD].longestPrefix() < KHMER_PREFIX_COMBINE_THRESHOLD)) {
                     // Look for a plausible word boundary
                     int remaining = rangeEnd - (current + wordLength);
                     int pc = fIter.current();
@@ -173,7 +175,7 @@ class LaoBreakEngine extends DictionaryBreakEngine {
                         }
                         if (fEndWordSet.contains(pc) && fBeginWordSet.contains(uc)) {
                             // Maybe. See if it's in the dictionary.
-                            int candidate = words[(wordsFound + 1) %LAO_LOOKAHEAD].candidates(fIter, fDictionary, rangeEnd);
+                            int candidate = words[(wordsFound + 1) %KHMER_LOOKAHEAD].candidates(fIter, fDictionary, rangeEnd);
                             fIter.setIndex(current + wordLength + chars);
                             if (candidate > 0) {
                                 break;
@@ -206,7 +208,7 @@ class LaoBreakEngine extends DictionaryBreakEngine {
             // We do this in code rather than using a rule so that the heuristic
             // resynch continues to function. For example, one of the suffix characters
             // could be a typo in the middle of a word.
-            // NOT CURRENTLY APPLICABLE TO LAO
+            // NOT CURRENTLY APPLICABLE TO KHMER
 
             // Did we find a word on this iteration? If so, push it on the break stack
             if (wordLength > 0) {
