+++ icu4j/main/classes/core/src/com/ibm/icu/text/RuleBasedBreakIterator.java
@@ -26,6 +26,14 @@ import com.ibm.icu.impl.CharacterIteration;
 import com.ibm.icu.impl.ICUBinary;
 import com.ibm.icu.impl.ICUDebug;
 import com.ibm.icu.impl.RBBIDataWrapper;
+import com.ibm.icu.impl.breakiter.BurmeseBreakEngine;
+import com.ibm.icu.impl.breakiter.CjkBreakEngine;
+import com.ibm.icu.impl.breakiter.DictionaryBreakEngine;
+import com.ibm.icu.impl.breakiter.KhmerBreakEngine;
+import com.ibm.icu.impl.breakiter.LanguageBreakEngine;
+import com.ibm.icu.impl.breakiter.LaoBreakEngine;
+import com.ibm.icu.impl.breakiter.ThaiBreakEngine;
+import com.ibm.icu.impl.breakiter.UnhandledBreakEngine;
 import com.ibm.icu.lang.UCharacter;
 import com.ibm.icu.lang.UProperty;
 import com.ibm.icu.lang.UScript;
